+++ core/domain/src/main/kotlin/com/metes/worthit/core/domain/entity/Item.kt
@@ -1,12 +1,13 @@
 package com.metes.worthit.core.domain.entity
 
+import java.math.BigDecimal
 import java.time.Instant
 import java.time.LocalDate
 
 data class Item(
     val id: Int = DEFAULT_ID,
     val name: String,
-    val price: Double?,
+    val price: BigDecimal?,
     val currency: Currency,
     val createdAt: Instant,
     val dateOfPurchase: LocalDate,
